app/frontend/requirements.txt
streamlit
requests
python-dotenv
extra-streamlit-components
